chore: fix hydration and font issues, refractor navbar
