+++ package.json
@@ -21,7 +21,7 @@
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "tailwind-merge": "^3.3.0",
-    "validator": "^13.15.0"
+    "validator": "^13.15.0",
     "react-icons": "^5.5.0",
     "zustand": "^5.0.5"
   },
